+++ tests/test_rebase_migration.py
@@ -14,7 +14,7 @@ from django.test import SimpleTestCase, TestCase, override_settings
 module = import_module("django_linear_migrations.management.commands.rebase-migration")
 
 
-class MakeMigrationsTests(TestCase):
+class RebaseMigrationsTests(TestCase):
     @pytest.fixture(autouse=True)
     def tmp_path_fixture(self, tmp_path):
         migrations_module_name = "migrations" + str(time.time()).replace(".", "")
@@ -238,6 +238,43 @@ class MakeMigrationsTests(TestCase):
             "Encountered a SyntaxError trying to parse 'dependencies = [(]'."
         )
 
+    def test_error_for_no_dependencies(self):
+        (self.migrations_dir / "__init__.py").touch()
+        (self.migrations_dir / "0001_initial.py").touch()
+        (self.migrations_dir / "0002_author_nicknames.py").touch()
+        (self.migrations_dir / "0002_longer_titles.py").write_text(
+            dedent(
+                """\
+            from django.db import migrations
+
+            class Migration(migrations.Migration):
+                dependencies = [
+                    ("otherapp", "0001_initial"),
+                ]
+                operations = []
+            """
+            )
+        )
+        (self.migrations_dir / "max_migration.txt").write_text(
+            dedent(
+                """\
+            <<<<<<< HEAD
+            0002_author_nicknames
+            =======
+            0002_longer_titles
+            >>>>>>> 123456789 (Increase Book title length)
+            """
+            )
+        )
+
+        with pytest.raises(CommandError) as excinfo:
+            self.call_command("testapp")
+
+        assert excinfo.value.args[0] == (
+            "Cannot edit '0002_longer_titles.py' since it has 0 dependencies"
+            + " within testapp."
+        )
+
     def test_error_for_double_dependencies(self):
         (self.migrations_dir / "__init__.py").touch()
         (self.migrations_dir / "0001_initial.py").touch()
@@ -272,7 +309,7 @@ class MakeMigrationsTests(TestCase):
             self.call_command("testapp")
 
         assert excinfo.value.args[0] == (
-            "Cannot edit '0002_longer_titles.py' since it has two dependencies"
+            "Cannot edit '0002_longer_titles.py' since it has 2 dependencies"
             + " within testapp."
         )
 
@@ -321,6 +358,64 @@ class MakeMigrationsTests(TestCase):
         assert not (self.migrations_dir / "0002_longer_titles.py").exists()
         new_content = (self.migrations_dir / "0003_longer_titles.py").read_text()
         deps = "[('testapp', '0002_author_nicknames'), ('otherapp', '0001_initial')]"
+        assert new_content == dedent(
+            f"""\
+            from django.db import migrations
+
+            class Migration(migrations.Migration):
+                dependencies = {deps}
+                operations = []
+            """
+        )
+
+    def test_success_swappable_dependency(self):
+        (self.migrations_dir / "__init__.py").touch()
+        (self.migrations_dir / "0001_initial.py").touch()
+        (self.migrations_dir / "0002_longer_titles.py").write_text(
+            dedent(
+                """\
+            from django.db import migrations
+
+            class Migration(migrations.Migration):
+                dependencies = [
+                    ('testapp', '0001_initial'),
+                    migrations.swappable_dependency('otherapp.0001_initial'),
+                ]
+                operations = []
+            """
+            )
+        )
+        (self.migrations_dir / "0002_author_nicknames.py").touch()
+        max_migration_txt = self.migrations_dir / "max_migration.txt"
+        max_migration_txt.write_text(
+            dedent(
+                """\
+            <<<<<<< HEAD
+            0002_author_nicknames
+            =======
+            0002_longer_titles
+            >>>>>>> 123456789 (Increase Book title length)
+            """
+            )
+        )
+
+        out, err, returncode = self.call_command("testapp")
+
+        assert out == (
+            "Renamed 0002_longer_titles.py to 0003_longer_titles.py,"
+            + " updated its dependencies, and updated max_migration.txt.\n"
+        )
+        assert err == ""
+        assert returncode == 0
+        max_migration_txt = self.migrations_dir / "max_migration.txt"
+        assert max_migration_txt.read_text() == "0003_longer_titles\n"
+
+        assert not (self.migrations_dir / "0002_longer_titles.py").exists()
+        new_content = (self.migrations_dir / "0003_longer_titles.py").read_text()
+        deps = (
+            "[('testapp', '0002_author_nicknames'), "
+            + "migrations.swappable_dependency('otherapp.0001_initial')]"
+        )
         assert new_content == dedent(
             f"""\
             from django.db import migrations